refactor(pdf-export): extract Supabase fetch + image resolution into pdf/data
